nirvana-reborn 2021.05.07 1、管理平台用户登录，获取登录验证码，功能开发完成 2、用户权限菜单界面与前端vue权限对接开发完成 3、用户列表页面分页查询，模糊查询，用户新增，用户删除功能开发完成 4、SysUserDO 对象新增数据库字段 sex：性别，email：邮箱，修改deptId字段为departmentId 5、HttpMessageConverter 配置全局web参数，long类型字段返回前端精度缺失问题修改 6、根据当前用户权限查询所能看到的部门数据 7、对分页查询与分页数据返回新增抽象类 8、整合了MD5加密工具类等能够用到的公共类等

+++ common/common-framework/src/main/java/com/phoenix/nirvana/common/vo/PageResult.java
@@ -24,7 +24,7 @@ public class PageResult<T> implements Serializable {
     private Long totalPage;
 
     @ApiModelProperty("数据总数")
-    private Long totalCount;
+    private Integer totalCount;
 
     @ApiModelProperty("分页数据")
     private List<T> list;
@@ -38,6 +38,6 @@ public class PageResult<T> implements Serializable {
                 .setList(Collections.EMPTY_LIST)
                 .setTotalPage(0L)
                 .setPageNo(0)
-                .setTotalCount(0L);
+                .setTotalCount(0);
     }
 }
